Rename ImageProcessorSrc to src
