--- 03/ImageProcessorSrc/cache/cache.cpp
@@ -1,3 +1,0 @@
-export module ips.cache;
-
-export import :loader;
