--- 03/ImageProcessorSrc/command/command.cpp
@@ -1,4 +1,0 @@
-export module ips.command;
-
-export import :parser;
-export import :argument;
