--- 03/ImageProcessorSrc/image/image.cpp
@@ -1,3 +1,0 @@
-export module ips.image;
-
-export import :processor;
