--- 03/ImageProcessorSrc/network/network.cpp
@@ -1,6 +1,0 @@
-export module ips.network;
-
-export import :server;
-export import :request;
-export import :response;
-export import :connection;
