+++ 03/src/cache/cache.cpp
@@ -1,0 +1,3 @@
+export module ips.cache;
+
+export import :loader;
