+++ 03/src/command/command.cpp
@@ -1,0 +1,4 @@
+export module ips.command;
+
+export import :parser;
+export import :argument;
