+++ 03/src/image/image.cpp
@@ -1,0 +1,3 @@
+export module ips.image;
+
+export import :processor;
